adapted to moved visitor classes
