+++ bundles/org.palladiosimulator.reliability.solver/src/org/palladiosimulator/reliability/solver/pcm2markov/MarkovTransformation.java
@@ -13,7 +13,7 @@ import org.palladiosimulator.reliability.solver.MarkovSolver;
 import org.palladiosimulator.reliability.solver.sensitivity.MarkovSensitivity;
 import org.palladiosimulator.solver.core.models.PCMInstance;
 import org.palladiosimulator.solver.core.runconfig.PCMSolverWorkflowRunConfiguration;
-import org.palladiosimulator.solver.visitors.UsageModelVisitor;
+import org.palladiosimulator.solver.core.visitors.UsageModelVisitor;
 import org.palladiosimulator.pcm.usagemodel.UsageScenario;
 
 /**
