+++ bundles/org.palladiosimulator.reliability.solver/src/org/palladiosimulator/reliability/solver/pcm2markov/MarkovTransformationSource.java
@@ -13,7 +13,7 @@ import org.palladiosimulator.pcm.resourceenvironment.ResourceContainer;
 import org.palladiosimulator.pcm.resourcetype.ProcessingResourceType;
 import org.palladiosimulator.pcm.seff.seff_performance.ParametricResourceDemand;
 import org.palladiosimulator.solver.core.models.PCMInstance;
-import org.palladiosimulator.solver.transformations.ContextWrapper;
+import org.palladiosimulator.solver.core.transformations.ContextWrapper;
 
 /**
  * Holds state information required during the PCM2Markov transformation.
