+++ bundles/org.palladiosimulator.reliability.solver/src/org/palladiosimulator/reliability/solver/pcm2markov/MarkovUsageModelVisitor.java
@@ -9,9 +9,9 @@ import org.palladiosimulator.reliability.MarkovEvaluationType;
 import org.palladiosimulator.reliability.markov.MarkovChain;
 import org.palladiosimulator.reliability.markov.State;
 import org.palladiosimulator.reliability.markov.StateType;
-import org.palladiosimulator.solver.transformations.ContextWrapper;
-import org.palladiosimulator.solver.utils.ManagedPMFParser;
-import org.palladiosimulator.solver.visitors.EMFQueryHelper;
+import org.palladiosimulator.solver.core.transformations.ContextWrapper;
+import org.palladiosimulator.solver.core.utils.ManagedPMFParser;
+import org.palladiosimulator.solver.core.visitors.EMFQueryHelper;
 import org.palladiosimulator.pcm.seff.ServiceEffectSpecification;
 import org.palladiosimulator.pcm.usagemodel.AbstractUserAction;
 import org.palladiosimulator.pcm.usagemodel.Branch;
